FEATURE - adding tick and improve text render

+++ src/main/java/dev/gustavo/components/entities/interfaces/IEntity.java
@@ -5,5 +5,5 @@ import org.joml.Matrix4f;
 public interface IEntity {
     Matrix4f getModelMatrix();
 
-    void update(final long delta);
+    void update(final float delta);
 }
